# BiuroPodrozy.Data/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BiuroPodrozyAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Languages": {
      "get": {
        "tags": [
          "Languages"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Languages"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      }
    },
    "/api/Languages/{id}": {
      "get": {
        "tags": [
          "Languages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Languages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Languages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pages": {
      "get": {
        "tags": [
          "Pages"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pages"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pages/{id}": {
      "get": {
        "tags": [
          "Pages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Page"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Sites": {
      "get": {
        "tags": [
          "Sites"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sites"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              }
            }
          }
        }
      }
    },
    "/api/Sites/{id}": {
      "get": {
        "tags": [
          "Sites"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sites"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Site"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Sites"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "City": {
        "type": "object",
        "properties": {
          "whoCreateId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModificationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deletionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "nullable": true
          },
          "auditDescription": {
            "type": "string",
            "nullable": true
          },
          "idCity": {
            "type": "integer",
            "format": "int32"
          },
          "cityName": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "pattern": "^[a-zA-Z0-9-]+$",
            "type": "string",
            "nullable": true
          },
          "population": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCountry": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            },
            "nullable": true
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            },
            "nullable": true
          },
          "flightsDep": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Flight"
            },
            "nullable": true
          },
          "flightsArr": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Flight"
            },
            "nullable": true
          },
          "trips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trip"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Country": {
        "type": "object",
        "properties": {
          "whoCreateId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModificationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deletionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "nullable": true
          },
          "auditDescription": {
            "type": "string",
            "nullable": true
          },
          "idCountry": {
            "type": "integer",
            "format": "int32"
          },
          "countryName": {
            "type": "string",
            "nullable": true
          },
          "abbrCountryName": {
            "maxLength": 3,
            "type": "string",
            "nullable": true
          },
          "continent": {
            "type": "string",
            "nullable": true
          },
          "population": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "countrySize": {
            "minimum": 0,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "phoneCode": {
            "pattern": "^[0-9]{2,3}$",
            "type": "string",
            "nullable": true
          },
          "idCurrency": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "idLanguage": {
            "type": "integer",
            "format": "int32"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "idCapitalCity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "capitalCity": {
            "$ref": "#/components/schemas/City"
          },
          "cities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/City"
            },
            "nullable": true
          },
          "countryPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountryPhoto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Currency": {
        "type": "object",
        "properties": {
          "whoCreateId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModificationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deletionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "nullable": true
          },
          "auditDescription": {
            "type": "string",
            "nullable": true
          },
          "idCurrency": {
            "type": "integer",
            "format": "int32"
          },
          "currencyName": {
            "type": "string",
            "nullable": true
          },
          "currencySymbol": {
            "type": "string",
            "nullable": true
          },
          "abbrCurrency": {
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "countries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Country"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Language": {
        "type": "object",
        "properties": {
          "whoCreateId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModificationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deletionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "nullable": true
          },
          "auditDescription": {
            "type": "string",
            "nullable": true
          },
          "idLanguage": {
            "type": "integer",
            "format": "int32"
          },
          "languageName": {
            "type": "string",
            "nullable": true
          },
          "abbrLanguageName": {
            "type": "string",
            "nullable": true
          },
          "countries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Country"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Page": {
        "type": "object",
        "properties": {
          "whoCreateId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModificationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deletionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "nullable": true
          },
          "auditDescription": {
            "type": "string",
            "nullable": true
          },
          "idPage": {
            "type": "integer",
            "format": "int32"
          },
          "pageTitle": {
            "type": "string",
            "nullable": true
          },
          "navTitle": {
            "type": "string",
            "nullable": true
          },
          "pageIndexRelativeURL": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "accessLevel": {
            "type": "string",
            "nullable": true
          },
          "includeInNavbar": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Site": {
        "type": "object",
        "properties": {
          "whoCreateId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModificationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastModificationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deletionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "nullable": true
          },
          "auditDescription": {
            "type": "string",
            "nullable": true
          },
          "idSite": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "tag": {
            "type": "string",
            "nullable": true
          },
          "logo": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}